# public/api.yaml
openapi: 3.0.3
info:
  title: Cluster Integrations API
  version: 1.0.0

servers:
  - url: https://cluster-main-eewh0y.laravel.cloud/api/v1

tags:
  - name: generic
    description: A generic endpoint
    externalDocs:
      description: Find out more
      url: http://swagger.io
  - name: authentication
    description: Authentication endpoints

paths:
  /integrations/generic:
    post:
      tags:
        - generic
      summary: Creates a new record of an entity
      description: Add a new record of an entity
      operationId: addRecord
      requestBody:
        description: Create a new record of an entity
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Data'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Invalid input
        '422':
          description: Validation exception

  /authenticate:
    post:
      tags:
        - authentication
      summary: Authenticate user
      description: Generates an authentication token for API access
      operationId: authenticateUser
      requestBody:
        description: User login credentials
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
        required: true
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Unauthorized

  /user:
    get:
      tags:
        - authentication
      summary: Get authenticated user details
      description: Fetch the details of the currently authenticated user
      operationId: getUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized

  /users:
    get:
      tags:
        - authentication
      summary: Get list of users
      description: Retrieve a list of all users
      operationId: getUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          description: Unauthorized

components:
  schemas:
    Data:
      required:
        - data
        - entity
      type: object
      properties:
        data:
          type: object
          example:
            id: "test001"
            first_name: "Jane"
            last_name: "Doe"
        entity:
          type: string
          example: Contact

    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          example: 200
        message:
          type: string
          example: success

    User:
      type: object
      properties:
        id:
          type: string
          example: "123"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          example: "johndoe@example.com"

    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKey:
      type: apiKey
      name: X-API-KEY
      in: header

security:
  - bearerAuth: []
  - ApiKey: []